refactoring filter/sorter
